% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawaname.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Draw_a_name}
\usage{
draw_a_name(pren = "Vincent", s = "M")
}
\arguments{
\item{pren}{Le prénom en question}

\item{s}{Le genre du prénom}
}
\value{
un graphique ggplot du prénom
}
\description{
Dessine un graphique du nombre de fois que le prénom a été donné par année
}
\examples{
\dontrun{
draw_a_name()
draw_a_name("Gustave","M")
draw_a_name("Marie","F")
}
}
