% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annéebissextile.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Is_leap_year}
\usage{
is_leap_year(x)
}
\arguments{
\item{x}{une année}
}
\value{
une belle phrase disant si l'année est bissextile
}
\description{
Détermine si une année est bissextile
}
\examples{
\dontrun{
is_leap_year(2013)
}
}
