% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiexcel.R
\name{multi_excel}
\alias{multi_excel}
\title{Multi_excel}
\usage{
multi_excel(path)
}
\arguments{
\item{path}{le chemin qui mene au fichier excel}
}
\value{
a list of dataframes
}
\description{
Lis un fichier excel (xlsx)
}
\examples{
\dontrun{
my_csv_reader("fichier.xlsx")
}
}
