% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycsvreader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{my_csv_reader}
\usage{
my_csv_reader(folder = getwd())
}
\arguments{
\item{folder}{un dossier}
}
\value{
(ouvre le fichier)
}
\description{
Lit des fichiers au format ".csv"
}
\examples{
my_csv_reader(fichier.csv)
}
