% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mygoodsum.R
\name{my_good_sum}
\alias{my_good_sum}
\title{My_good_sum}
\usage{
my_good_sum(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{another number}
}
\value{
the sum
}
\description{
This computes the sum of two numbers
}
\examples{
\dontrun{
my_good_sum(3,4)
}
}
