% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{print_pokemon}
\alias{print_pokemon}
\title{Print_pokemon}
\usage{
print_pokemon(x)
}
\arguments{
\item{x}{un pokemon}
}
\description{
Reconnait et présente un pokemon
}
\examples{
\dontrun{
print.pokemon("Dracaufeu")
}
}
