% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveascsv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save_as_csv}
\usage{
save_as_csv(dataset, filename, row.names = FALSE, ...)
}
\arguments{
\item{x}{un jeu de données, ou dataset}

\item{y}{un nom de fichier}
}
\description{
Cette fonction enregistre un jeu de données au format csv
}
\examples{
\dontrun{
save_as_csv(fichier)
}
}
