% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveascsv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save_as_csv}
\usage{
save_as_csv(dataset, filename, row.names = FALSE, ...)
}
\arguments{
\item{dataset}{un jeu de données}

\item{filename}{un nom de fichier}

\item{row.names}{des noms de lignes}
}
\description{
Cette fonction enregistre un jeu de données au format csv
}
\examples{
\dontrun{
save_as_csv(fichier)
}
}
